% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_IT_segments.R
\name{clip_IT_segments}
\alias{clip_IT_segments}
\title{Clip individual tree segments from a point cloud using crown polygons}
\usage{
clip_IT_segments(
  las_dir_path,
  crowns_path,
  out_path,
  treeID = "treeID",
  defined_crs = 2193,
  workers = 10
)
}
\arguments{
\item{las_dir_path}{Character. Path to the directory containing LAS files.}

\item{crowns_path}{Character. Path to the shapefile containing tree crown
polygons.Crowns must have a unique ID, see below.}

\item{out_path}{Character. Path to the directory where output files will be
saved.A new directory will be created here.}

\item{treeID}{Character. Name of the column in crowns file that uniquely
identifies each tree.}

\item{defined_crs}{Numeric. Coordinate reference system (CRS) for output
layers.
Default is EPSG:2193.}

\item{workers}{Integer. Number of workers/logical processors to use for
processing. This is used to set up parallel processing.
Default is 10.}
}
\value{
Clipped individual tree segments will be saved in a newly created
        directory.A list containing paths to these segments will also be
        returned.
}
\description{
This function reads tree crown polygons and then clips individual tree
segments based on the provided polygons. Each tree segment is saved as an
individual LAS file.
}
\details{
This function performs the following steps:
\itemize{
  \item Reads the tree crown polygons from the provided path.
  \item Reads the LAS files as a catalog.
  \item Creates an output directory to store the clipped tree segments.
  \item Clips individual tree segments from the LAS files, saving each tree
        segment as a separate file.
  \item Returns the file paths to the clipped LAS files, each representing an
        individual tree segment.
}
}
\examples{
\dontrun{
clip_IT_segments(
  las_dir_path = "path/to/las/files",
  crowns_path = "path/to/crowns.shp",
  out_path = "path/to/output")
}

}
