% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_crowns.R
\name{delineate_crowns}
\alias{delineate_crowns}
\title{Delineate tree crowns on the canopy height model (CHM)}
\arguments{
\item{chm_path}{Character. Path to the CHM raster file.}

\item{stem_map_path}{Character. Path to the stem map (point shapefile).}

\item{min_ht}{Numeric. Minimum height threshold in meters for crown
delineation. Any pixels having height below this value will be
excluded from the crown delineation process.}

\item{smooth}{Logical. If \code{TRUE}, smooths the edges of delineated
crowns.
Default is \code{TRUE}.}

\item{defined_crs}{Numeric. Coordinate reference system (CRS) for output
layers.
Default is EPSG:2193.}
}
\value{
A polygon object containing the delineated crown polygons with their
        matched stem IDs.
}
\description{
This function delineates individual tree crowns from a canopy height model
(CHM) raster using a watershed segmentation approach. The resulting polygons
represent crown boundaries and are matched to the provided stem map.
}
\details{
This function performs the following steps:
\itemize{
  \item Reads and processes the CHM raster, applying focal filters to fill
        gaps and smooth raster).
  \item Uses watershed segmentation to delineate tree crowns using stem
        map locations of trees as seed points.
  \item Label the delineated crowns with matching stem ID.
  \item Optionally (if specified) smooths the edges of the crown polygons.
}
}
\examples{
\dontrun{
segs_sf <- delineate_crowns(
  chm_path = "path/to/chm.tif",
  stem_map_path = "path/to/stem_map.shp")
ggplot2::ggplot(data = crowns) +
  ggplot2::geom_sf() +
  ggplot2::geom_sf_text(ggplot2::aes(label = treeID),
                        size = 3, color = "blue") +
  ggplot2::theme_bw() +
  ggplot2::labs(title = "Crowns with tree IDs")
}

}
