% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_stem.R
\name{detect_stem}
\alias{detect_stem}
\title{Stem detection from LAS individual tree point cloud}
\arguments{
\item{las_path}{Character. The file path to the LAS file.}

\item{stem_only}{Logical. If \code{TRUE}, returns only stem points if
\code{FALSE}, returns all points classified as Stem and
non-Stem.
Defaults to \code{TRUE}.}

\item{min_ht}{Numeric. The minimum height threshold in meters for filtering
stem points. Points below this height will be excluded.}

\item{max_ht}{Numeric or character. The maximum height limit in meters for
filtering stem points, either as "zmax" to use the file's
maximum Z value or a specific numeric height.
Default is \code{"zmax"}.}

\item{all_attr}{Logical. If \code{TRUE}, reads all attributes from the LAS
file and keep them through the processing if \code{FALSE},
reads and retains only XYZ coordinates.
Defaults to \code{TRUE}.}

\item{ht_step}{Numeric. Height step interval (in meters) used in the Hough
transform for detecting stem segments. A smaller value
increases accuracy but may increase computation time.
Default is 0.1.}

\item{max_diameter}{Numeric. Maximum allowable stem diameter in meters
for classification. Points representing stems larger than
this diameter will be excluded.
Default is 0.6.}

\item{ht_base}{Numeric vector of length 2. Specifies the height range
(in meters) considered as the base of the tree for stem
detection.
Default is \code{c(1, 2)}.}

\item{pixel_res}{Numeric. Pixel resolution (in meters) for grid-based stem
classification.
Default is 0.02 meters.}

\item{min_den}{Numeric. Minimum density of points (per square meter) required
for an area to be considered part of the stem.
Default is 0.2.}

\item{min_numvotes}{Integer. Minimum number of votes in the Hough transform
required to confirm stem points. Lower values may
increase false positives.
Default is 3.}

\item{defined_crs}{Numeric. Coordinate reference system (CRS) for the output
LAS object.
Default is EPSG:2193.}
}
\value{
A LAS object containing detected stem and non-stem (if specified)
        points that meet the specified height and classification criteria.
}
\description{
This function detects and classifies stem points in an individual tree
LAS file, allowing restriction of the height range based on user input.
It uses the Hough transformation method to identify and classify the stem
cylinder in the point cloud.
}
\details{
This function performs stem detection on a LAS file by:
\itemize{
  \item Reads the LAS file and assigns the specified CRS.
  \item Checks for missing values in the X, Y, and Z columns.
  \item Filters points based on the height range defined by \code{min_ht}
        and \code{max_ht}.
  \item Classify stem points using the Hough transformation, restricting
        results based on parameters like \code{max_diameter}, \code{ht_step},
        and \code{min_numvotes}.
  \item Returns a LAS object containing only the detected stem points
        meeting these criteria.
}
}
\examples{
\dontrun{
stem_points <- detect_stem(las_path = "path/to/your/file.laz",
   stem_only = TRUE)
lidR::plot(stem_points)
}

}
