% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_ht_diameter_DF.R
\name{fixedht_diameter_DF}
\alias{fixedht_diameter_DF}
\title{Calculate diameter at a specified fixed height from a diameter profile}
\arguments{
\item{diameter_profile}{A dataframe containing the diameter profile of a tree
with columns `z` (height) and `diameter`.}

\item{fixed_ht}{Numeric. The fixed height (in meters) at which to retrieve
the diameter.
Default is 1.4 meters.}
}
\value{
A dataframe with columns: `x`, `y`, `z`, `n_points`, and `diameter`
        (in meters), representing the closest diameter at the specified
        height.
}
\description{
This function retrieves the estimated diameter of a tree stem at a specified
fixed height from a pre-computed diameter profile.
}
\details{
This function identifies the closest diameter to a specified height by:
\itemize{
  \item Calculates distance between each \code{z} value in
        \code{diameter_profile} and the specified \code{fixed_ht} to
        determine proximity.
  \item Selects the smallest absolute difference, which represents the
        diameter measurement closest to \code{fixed_ht}.
  \item Returns a dataframe with the relevant details, including: `x`, `y`,
        `z`, `n_points`, and `diameter` (in meters).
}
}
\examples{
\dontrun{
stem_points <-  detect_stem(las_path = "path/to/your/file.laz",
   stem_only = TRUE)
d_profile <- stem_diameter_profile(stem_points)
diameter_at_fixed_ht <- fixedht_diameter_DF(d_profile,
   fixed_ht = 1.4)
}

}
