% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_ht_diameter_LAS.R
\name{fixedht_diameter_LAS}
\alias{fixedht_diameter_LAS}
\title{Calculate diameter at a specified fixed height from a LAS object}
\arguments{
\item{stem_las}{A LAS object representing a tree stem point cloud.}

\item{fixed_ht}{Numeric. The fixed height (in meters) at which to calculate
the diameter.
Default is 1.4.}

\item{seg_ht}{Numeric. The vertical interval (in meters) above and below
\code{fixed_ht} used to define the segment height range.
Default is 0.2.}

\item{n_points}{Integer. Minimum number of points required in the segment
for diameter estimation.
Default is 25.}

\item{quantiles}{Numeric vector of length 2. Quantiles for filtering points
within the segment based on their distance to the shape
center. The values must be between 0 and 1.
Default is \code{c(0.1, 0.9)}.}
}
\value{
A tibble with columns: x, y, z, n_points, and diameter (in meters)
        representing the calculated values at the specified height.
}
\description{
This function calculates the diameter of a tree stem at a specified fixed
height from an individual tree point cloud.
}
\details{
This function calculates the diameter of a tree stem at a specified fixed
height by performing the following steps:
\itemize{
  \item Extracts a vertical segment of points from the point cloud within a
        range of \code{fixed_ht} ± \code{seg_ht}/2 meters.
  \item Filters points within this segment based on their distances from the
        estimated shape center, using quantiles specified in
        \code{quantiles} to remove outliers.
  \item Fits a shape to the filtered points to estimate the diameter at the
        specified height.
  \item Returns a dataframe containing the fitted shape's \code{x} and
        \code{y} center coordinates, the specified \code{z} height, the count
        of points \code{n_points} used, and the calculated \code{diameter}.
}
}
\examples{
\dontrun{
stem_points <-  detect_stem(las_path = "path/to/your/file.laz",
    stem_only = TRUE)
diameter_at_fixed_ht <- fixedht_diameter_LAS(stem_points,
    fixed_ht = 1.4,
    seg_ht = 0.25)
}

}
