% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dbh_ht.R
\name{get_dbh_ht}
\alias{get_dbh_ht}
\title{Detect the nearest internode height where DBH must be measured
(PlotSafe Guideline Replication)}
\arguments{
\item{node_profile}{A dataframe containing the node profile with columns
`z_min`, `z_max`, and `section_type`, where
`section_type` denotes "Node" or "Internode".}

\item{bh}{Numeric. The target height in meters to find the closest internode.
Default is 1.4 meters.}

\item{min_ht}{Numeric. The minimum height in meters to consider for internode
detection.
Default is 0.2 meters.}

\item{max_ht}{Numeric. The maximum height in meters to consider for internode
detection.
Default is 2.9 meters.}

\item{tol1}{Numeric. Tolerance range (in meters) around the target height for
a direct internode measurement.
Default is 0.1.}

\item{tol2}{Numeric. Additional tolerance range (in meters) for
finding overlapping internode segments if no direct measurement
is possible and split is required.
Default is 0.3.}
}
\value{
A tibble with columns:`dbh_class`, which is the classification of the
       DBH measurement type ("DirectM", "DirectSS", "DirectSS", or "Split"),
       list, and `nearest_int2`, an additional internode height or range if a
       split measurement is needed.
\itemize{
  \item DirectM: Diameter is measured at breast height (1.4 meters) directly
        over an internode.
  \item Direct SS: Diameter is measured directly at a distance of `tol1`
        above or below breast height, over an internode (1.3 to 1.5 m).
  \item Split: To avoid a long welling, diameter measurements are taken at
        two locations (internodes) over the internodes and then averaged.
  \item DirectLS: When diameter cannot be measured using the above methods
        due to the presence of a long node or swollen area, a direct
        measurement is taken between 0.2 and 2.9 meters over an internode.
}
}
\description{
This function identifies the nearest internode in a stem profile where DBH
must be measured according to the PlotSafe Guidelines. Typically breast
height is 1.4 meters for direct measurements. If no direct measurement is
possible, this function estimates the nearest height to 1.4 meters where no
nodes/swellings are present.
}
\details{
This function detects the nearest internode height to a specified target
height, typically used to find the closest internode around breast height
(1.4 meters). The function operates as follows:
\itemize{
  \item  Attempts to identify an internode that overlaps with the specified
         height (`bh`) within a small tolerance (`tol1`).
  \item If no direct measurement is available, the function searches for a
        pair of internode sections, one above and one below the target
        height, to calculate an estimated height.
  \item If neither a direct nor split measurement is possible, the function
        finds the closest available internode below the target height.
}
}
\examples{
\dontrun{
stem_points <- detect_stem(las_path = "path/to/your/file.laz",
    stem_only = TRUE)
d_profile <- stem_diameter_profile(stem_points)
n_profile <- stem_node_profile(d_profile, tol = 1, base = 4,
    split = c(FALSE, "Node"))
dbh_ht <- variable_DBH_ht(n_profile, bh = 1.4, min_ht = 0.2,
    max_ht = 2.9, tol1 = 0.1, tol2 = 0.3)
}

}
