% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_crown_metrics_ALS.R
\name{get_tree_metrics_ALS}
\alias{get_tree_metrics_ALS}
\title{Calculate height and crown metrics from an individual tree point cloud}
\arguments{
\item{las_path}{Character. The file path to the LAS file.}

\item{height}{Logical. If \code{TRUE}, calculates height metrics.
Default is \code{TRUE}.}

\item{crown}{Logical. If \code{TRUE}, calculates crown metrics.
Default is \code{TRUE}.}

\item{min_ht}{Numeric. The minimum height (in meters) to filter points.
Points with heights below this will be excluded.
Default is 0.3.}

\item{defined_crs}{Integer. Coordinate Reference System (CRS) to assign to
the LAS point cloud.
Defaults is EPSG:2193.}
}
\value{
A dataframe with calculated height metrics (if specified) and crown
        metrics (if specified).
}
\description{
This function calculates height and crown metrics from a point cloud.
This is specifically for extracting metrics from airborne laser scanning
(ALS) data, including unmanned aerial vehicle laser scanning (ULS).
Users can toggle height and crown metric calculations separately and apply
a minimum height filter to exclude points below a specified height.
}
\examples{
\dontrun{
las <- readLAS("path/to/lasfile.las")
metrics_ALS <- get_tree_metrics(las_path,
    height = TRUE,
    crown = TRUE,
    min_ht = 2)
print(metrics_ALS)
}

}
