% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_crown_metrics_CHM.R
\name{get_tree_metrics_CHM}
\alias{get_tree_metrics_CHM}
\title{Calculate height and crown metrics from CHM}
\arguments{
\item{chm_path}{Character. The file path to the canopy height model (CHM)
raster.}

\item{crowns_path}{Character. The file path to the crown polygons.}

\item{height}{Logical. If \code{TRUE}, calculates height metrics.
Default is \code{TRUE}.}

\item{crown}{Logical. If \code{TRUE}, calculates crown area.
Default is \code{TRUE}.}

\item{min_ht}{Numeric. The minimum height threshold (in meters) for
filtering pixels in CHM. Pixels with heights below this will be
excluded from metric calculations.
Default is 2.}

\item{treeID}{Character. Name of the column in the crowns data that
identifies individual trees.
Default is "treeID".}

\item{defined_crs}{Integer. EPSG code for the Coordinate Reference System to
be assigned to both the CHM and crowns.
Default is EPSG:2193.}
}
\value{
A dataframe with calculated height metrics (if specified) and crown
        area (if specified) for each crown polygon.
}
\description{
This function calculates height metrics for each crown polygon, excluding
pixels below a specified minimum height threshold. Optionally, it calculates
the crown area for each polygon.
}
\details{
\itemize{
  \item{Height metrics}{Calculates statistics including max, mean, standard
                        deviation, coefficient of variation, and specified
                        quantiles (5th, 50th, and 95th percentiles) for each
                        crown polygon, excluding pixels below the
                        \code{min_ht} threshold.}
  \item{Crown metrics}{Calculates the area of each crown polygon if
                       \code{crown = TRUE}.}
}
}
\examples{
\dontrun{
crown_metrics <- get_tree_metrics_CHM(
  chm_path = "path/to/chm.tif",
  crowns_path = "path/to/crowns.shp",
  height = TRUE,
  crown = TRUE,
  min_ht = 2,
  treeID = "treeID",
  defined_crs = 2193
)
print(crown_metrics)
}

}
