% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_crown_metrics_TLS.R
\name{get_tree_metrics_TLS}
\alias{get_tree_metrics_TLS}
\title{Calculate height and crown metrics from an individual tree point cloud}
\arguments{
\item{las_path}{Character. The file path to the LAS file.}

\item{height}{Logical. If \code{TRUE}, calculates height metrics.
Default is \code{TRUE}.}

\item{crown}{Logical. If \code{TRUE}, calculates crown metrics.
Default is \code{TRUE}.}

\item{min_ht}{Numeric. The minimum height (in meters) to filter points.
Points with heights below this will be excluded.
Default is 0.3.}

\item{defined_crs}{Integer. Coordinate Reference System (CRS) to assign to
the LAS point cloud.
Defaults is EPSG:2193.}
}
\value{
A dataframe with calculated height metrics (if specified) and crown
        metrics (if specified).
}
\description{
This function calculates height and crown metrics from a point cloud.
This is specifically for extracting metrics from mobile laser scanning (MLS)
or terrestrial laser scanning (TLS) data.
Users can toggle height and crown metric calculations separately and apply
a minimum height filter to exclude points below a specified height.
}
\details{
This function provides the following metrics:
\itemize{
  \item Height Metrics: If \code{height = TRUE}, the function calculates:
    \itemize{
      \item \code{maxHt}: Maximum height of the point cloud.
      \item \code{meanHt}: Mean height.
      \item \code{modeHt}: Median height.
      \item \code{stHt}: Standard deviation of heights.
      \item \code{CVHt}: Coefficient of variation of heights.
      \item Height percentiles: 5\%, 10\%, 20\%, 25\%, 30\%, 40\%, 50\%,
            60\%, 70\%, 75\%, 80\%, 90\%, 95\%, and 99\%.
    }

  \item Crown Metrics: If \code{crown = TRUE}, the function calculates:
    \itemize{
      \item \code{crownX}, \code{crownY}: Centroid location of the crown.
      \item \code{concavehull2D_area}: Area of the 2D concave hull.
      \item \code{convexhull2D_area}: Area of the 2D convex hull.
      \item \code{convexhull3D_volume}: Volume of the 3D convex hull.
      \item \code{convexhull3D_area}: Surface area of the 3D convex hull.
    }
    These metrics offer insights into canopy spread and crown volume.
}
}
\examples{
\dontrun{
path <- readLAS("path/to/lasfile.las")
metrics <- get_tree_metrics_TLS(las_path = path,
    height = TRUE,
    crown = TRUE,
    min_ht = 2)
print(metrics)
}

}
