% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenise_LAS.R
\name{homogenise_LAS}
\alias{homogenise_LAS}
\title{Decimate ultra-high-density point clouds to manageable size}
\usage{
homogenise_LAS(
  las_dir_path,
  out_path,
  point_density,
  grid_res = 0.1,
  workers = 10,
  defined_crs = 2193
)
}
\arguments{
\item{las_dir_path}{Character. Path to the directory containing LAS files.}

\item{out_path}{Character. Path to the directory where decimated LAS files
will be saved.}

\item{point_density}{Integer. Point density to be used in homogenisation.}

\item{grid_res}{Numeric. Grid resolution in meters. A 3D grid of the
specified size will be used for decimation.
Default is 0.1.}

\item{workers}{Integer. Number of workers/logical processors to use for
processing. This is used to set up parallel processing.
Default is 10.}

\item{defined_crs}{Numeric. Coordinate reference system (CRS) for the output
layers.
Default is EPSG:2193.}
}
\value{
A vector of file paths to the decimated LAS files.
}
\description{
This function reduces the density of high-resolution point clouds (e.g., from
MLS or TLS) to a manageable size suitable for further processing, such as
crown delineation. It performs spatial subsampling using a specified 3D grid
resolution.
}
\details{
This function performs the following steps:
\itemize{
  \item Reads the LAS files from the specified directory as a LAS catalog.
  \item Checks that the specified \code{num_points} is appropriate given the
        point density of the LAS catalog.
  \item Applies a decimation process using the specified grid resolution,
        \code{num_points} value and random point selection.
  \item Saves decimated LAS files in the specified output directory.
}
}
\examples{
\dontrun{
Decimate_LAS(
  las_dir_path = "path/to/las/files",
  out_path = "path/to/output"
)
}
}
