% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_diameter.R
\name{segment_diameter}
\alias{segment_diameter}
\title{Get diameter from a stem segment point cloud}
\arguments{
\item{segment}{A LAS object representing a segment of stem points.}

\item{quantiles}{Numeric vector of length 2. Quantiles for point filtering
within segments based on distance from the shape center.
Values must be between 0 and 1.
Default is \code{c(0.1, 0.9)}.}

\item{n_points}{Integer. Minimum number of points required per segment for
diameter estimation.
Default is 25.}
}
\value{
A tibble with columns: `x`, `y`, and `diameter` (in meters) if
        fitting is successful, or a tibble with `NA` values if fitting fails.
}
\description{
This function takes a segment of LAS points as input and fits a cylinder
to the points, filtering them based on specified distance quantiles and
calculating the shape parameters (center coordinates and diameter).
It returns `x`, `y` coordinates at the center and `diameter` if the fitting
is successful.
}
\details{
This function performs the following steps:
\enumerate{
  \item Verifies if the segment contains the minimum required number of
        points as specified by \code{n_points}.
  \item Attempts to fit a shape to the points in the segment using the
        RANSAC algorithm via \code{TreeLS::shapeFit()}.
  \item Filters points based on their distances to the initial shape center
        within the quantile range defined by \code{quantiles}.
  \item Refits a shape to the filtered points.
  \item If the refit is successful, returns the shape parameters (center
        \code{x}, \code{y}, and \code{diameter}). Otherwise, returns
        \code{NA} values.
}
}
\examples{
\dontrun{
segment <- readLAS("path/to/las_segment.laz")
diameter <- segment_diameter(segment,
   n_points = 30,
   quantiles = c(0.05, 0.95))
}

}
