% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_diameter_profile.R
\name{stem_diameter_profile}
\alias{stem_diameter_profile}
\title{Calculate stem diameter profile}
\arguments{
\item{stem_las}{A LAS object that has stem points detected and classified.}

\item{seg_ht}{Numeric. Height of each segment (profile interval) in meters.}

\item{min_ht}{Numeric. Minimum stem height for segmentation in meters.
Points below this height will be excluded. Default is 0.2.}

\item{max_ht}{Numeric or character. Maximum height for segmentation,
either as "zmax" (to use the maximum height in the LAS file) or
a specific numeric value. Default is \code{"zmax"}.}

\item{min_diameter}{Numeric. Minimum allowable stem diameter in meters.
Defaults to 0.05.}

\item{max_diameter}{Numeric. Maximum allowable stem diameter in meters.
Defaults to 0.6.}

\item{quantiles}{Numeric vector of length 2. Quantiles for point filtering
within each segment, used to exclude outliers in diameter
calculation.
Default is \code{c(0.1, 0.9)}.}

\item{n_points}{Integer. Minimum number of points per segment required for
diameter estimation. Default is 25.}
}
\value{
A dataframe with columns: `ID`, `x`, `y`, `z`, `points`, and
        `diameter`(in meters), representing the diameter profile of the stem.
}
\description{
This function calculates the diameter profile of a tree stem from an
individual tree point cloud.
}
\details{
This function calculates a diameter profile for a tree stem by performing the
following steps:
\itemize{
  \item Divides the point cloud into vertical segments, each with a height
        specified by \code{seg_ht}. Only points within the height
        range specified by \code{min_ht} and \code{max_ht} are used.
  \item Fits a shape to the filtered points in each segment and calculates
        the segment diameter. Only segments with at least \code{n_points}
        are considered for diameter estimation.
  \item Returns a data frame where each row represents a segment, including
        the estimated `x` and `y` center coordinates,`z` height (in meters),
        point count, and calculated `diameter` (in meters).
}
}
\examples{
\dontrun{
stem_points <-  detect_stem(las_path = "path/to/your/file.laz",
    stem_only = TRUE)
d_profile <- stem_diameter_profile(stem_points,
    seg_ht = 0.25,
    min_ht = 0.3,
    max_ht = "zmax",
    quantiles = c(0.2, 0.85),
    n_points = 25,
    max_diameter = 0.5,
    min_diameter = 0.1)
}

}
