% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_node_profile.R
\name{stem_node_profile}
\alias{stem_node_profile}
\title{Detect nodes and internodes from a diameter profile}
\arguments{
\item{diameter_profile}{A dataframe containing the diameter profile of a tree
with columns `z` (height) and `diameter` in meters.}

\item{tol}{Numeric. Tolerance threshold for identifying diameter changes in.
centimeters. Sections with diameter changes larger than this
threshold are classified as nodes, while sections with stable
diameter are classified as internodes.
Default is 1.}

\item{base}{Numeric. A base value used to inflate the initial segment
diameter to assist in node detection. The base will be multiplied
by tol and added to the diameter at the min_ht.
Default is 3.}

\item{split}{A vector with a logical value and a character string. The
character values must be either "Node" or "Internode". If the
logical value is \code{TRUE}, only sections of type specified in
the string (either "Node" or "Internode") are returned.
Default is \code{c(FALSE, "Node")}.}
}
\value{
A dataframe containing columns: `section_id`, `z_min`, `z_max`,
        `z_avg`, `section_length`, `diameter_lower`, `diameter_upper`,
        `diameter_avg` and `section_type`. The z values and diameters will be
        in meters.
}
\description{
This function analyzes the diameter profile of a tree stem to identify nodes
(areas of significant diameter increase) and internodes (areas of stability
or reduction in diameter). It provides detailed metrics for each identified
section, including height range and average diameter.
}
\details{
This function performs the following steps:
\itemize{
  \item Filters for continuous data, fills or drops missing values, and
        adjusts segment heights for consistent detection.
  \item Calculates the diameter difference for each segment and assigns a
        section type based on the tolerance threshold.
  \item Summarizes each section by height range, diameter range, average
        diameter, and section type.
  \item If specified in \code{split}, returns only sections of the specified
        type (either "Node" or "Internode").
}
}
\examples{
\dontrun{
stem_points <-  detect_stem(las_path = "path/to/your/file.laz",
    stem_only = TRUE)
d_profile <- stem_diameter_profile(stem_points)
n_profile <- node_profile(d_profile, tol = 1, base = 4,
    split = c(FALSE, "Node"))
}

}
