% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{stem_volume}
\alias{stem_volume}
\title{Calculate cumulative volume of a tree stem}
\arguments{
\item{diameter_profile}{A dataframe containing the diameter profile of a tree
with columns `z` (height) and `diameter`.}

\item{drop.NoData}{Logical. If \code{TRUE}, rows with missing diameter values
are omitted.
Default is \code{FALSE}.}

\item{fill.NoData}{A vector with a logical value and a character string. The
logical value indicates whether to fill missing diameters,
and the string specifies the fill direction ("up", "down",
or "updown"). `up` uses the preceding value, `down` uses
the next value, and `updown` fills as needed from both
directions.
Default is \code{c(TRUE, "updown")}.}
}
\value{
A dataframe with columns `vol_min_ht`, where the minimum height of
        the tree for which volume was calculated, `vol_max_ht`, where the
        maximum height of the tree for which volume was calculated, and
        `volume`, the cumulative volume (in cubic meters) of the tree stem
        within the height range.
}
\description{
This function calculates the cumulative volume of a tree stem based on its
diameter profile, using the formula for the volume of a cylinder for each
segment. Missing sectional diameter values can be optionally filled or
dropped.
}
\details{
This function performs the following steps:
\itemize{
  \item Checks that `diameter_profile` contains necessary columns and
        validates input parameters.
  \item Identifies the last valid height with data continuity and removes
        rows above this height.
  \item Determines the height of each segment based on consecutive `z`
        values.
  \item Fills or drops missing diameter values according to the user input.
  \item Calculates the volume for each segment and then computes the
        cumulative volume (in cubic meters) up to each height (in meters).
}
}
\examples{
\dontrun{
stem_points <-  detect_stem(las_path = "path/to/your/file.laz",
    stem_only = TRUE)
d_profile <- stem_diameter_profile(stem_points)
tree_volume <- stem_volume(d_profile)
}

}
