% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_character}
\alias{validate_character}
\title{Validate character input}
\usage{
validate_character(x)
}
\arguments{
\item{x}{Character. The value to validate.}
}
\value{
NULL if valid, otherwise throws an error.
}
\description{
Checks if a value is a non-null, non-empty character string.
}
