% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_crs}
\alias{validate_crs}
\title{Validate Coordinate Reference System (CRS) EPSG code}
\usage{
validate_crs(defined_crs)
}
\arguments{
\item{defined_crs}{Numeric. The EPSG code to validate.}
}
\value{
NULL if valid, otherwise throws an error.
}
\description{
Checks if a CRS code is numeric and within the EPSG code range.
}
