% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_dataframe}
\alias{validate_dataframe}
\title{Validate DataFrame input with specific columns}
\usage{
validate_dataframe(df, required_cols)
}
\arguments{
\item{df}{A data frame that is expected to contain certain columns.}

\item{required_cols}{A character vector of column names that the
DataFrame must have.}
}
\value{
NULL if valid, otherwise throws an error.
}
\description{
Checks if a given DataFrame contains the required columns.
}
