% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_las}
\alias{validate_las}
\title{Validate LAS object}
\usage{
validate_las(las, required_cols = c("X", "Y", "Z"))
}
\arguments{
\item{las}{An object expected to be of class `LAS` from the `lidR` package.}

\item{required_cols}{A vector containing all required column names.
Default is \code{c("X", "Y", "Z")}}
}
\value{
NULL if valid, otherwise throws an error.
}
\description{
Checks if the provided object is a valid `LAS` object from the `lidR`
package, is not empty, and contains necessary attributes.
}
