% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_logical}
\alias{validate_logical}
\title{Validate logical input}
\usage{
validate_logical(x)
}
\arguments{
\item{x}{The input value to be validated.}
}
\value{
NULL if valid, otherwise throws an error.
}
\description{
Checks if the provided input is a valid boolean/logical (TRUE or FALSE) value.
}
