% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_path}
\alias{validate_path}
\title{Validate various input types}
\arguments{
\item{path}{Character. The path to validate.}

\item{create}{Logical. If \code{TRUE}, attempts to create the directory if
given path does not exist
Default is FALSE.}
}
\value{
NULL if valid, otherwise throws an error.
}
\description{
Validate directory path
}
\details{
Checks if a path is a character string, if it exists, or if it can be
created.
}
