% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_vector}
\alias{validate_vector}
\title{Validate vector input}
\usage{
validate_vector(x, length, name)
}
\arguments{
\item{x}{The input value to be validated, expected to be a vector.}

\item{length}{The required length of the vector.}

\item{name}{(optional) The name of the variable, used for error messages.}
}
\value{
NULL if valid, otherwise throws an error.
}
\description{
Checks if the provided input is a vector of the specified length.
}
