% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_ht_dbh_DF.R
\name{variable_ht_dbh}
\alias{variable_ht_dbh}
\title{Estimate diameter at variable height}
\arguments{
\item{diameter_profile}{A dataframe containing the diameter profile of a tree
with columns `z` (height) and `diameter`.}

\item{var_ht}{A dataframe generated from `variable_DBH_ht` with columns
specifying breast height (`breast_ht`), tolerance
(`splitM_tol`), internode ranges, and DBH classification.}
}
\value{
A dataframe with columns: `dbh_class`, `dbh_height`, and `diameter`,
        representing the estimated diameter (in meters) based on height
        criteria. If the required height measurements or internode data are
        unavailable, the function returns \code{NA} values in the output
        dataframe.
}
\description{
This function estimates the diameter of a tree stem at variable heights from
a diameter profile, using on pre-computed internode information.
}
\details{
This function retrieves the estimated diameter at a height defined from a
specified diameter profile using internode ranges and t olerances:
\itemize{
  \item Calculates internode segment heights and nearest internode ranges.
  \item Determines the nearest height segments that overlap with the breast
        height (bh).
  \item Based on the `dbh_class`, selects the appropriate diameter measure:
        \code{"DirectM"} retrieves the diameter directly at breast height
        (1.4m), \code{"DirectSS"} finds the diameter closest to the nearest
        internode, and \code{"Split"} averages diameters within a split
        tolerance range.
}
}
\examples{
\dontrun{
stem_points <- detect_stem(las_path = "path/to/your/file.laz",
    stem_only = TRUE)
d_profile <- stem_diameter_profile(stem_points)
n_profile <- stem_node_profile(d_profile)
dbh_ht <- get_dbh_ht(n_profile)
dbh <- variable_ht_dbh(d_profile, dbh_ht)
print(dbh)
}

}
